+++ install.sh
@@ -6,7 +6,7 @@ else
   set -o xtrace
 fi
 
-VERSION="0.2.2"
+VERSION="0.3.0"
 
 if [ "$(uname)" == "Darwin" ]; then